% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-dependencies.R
\name{add_bootstrap_deps}
\alias{add_bootstrap_deps}
\title{bootstrap dependencies utils}
\usage{
add_bootstrap_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches bootstrap. dependencies to the given tag
}
