% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbdash-dependencies.R
\name{add_cbdash_deps}
\alias{add_cbdash_deps}
\title{cbdash dependencies utils}
\usage{
add_cbdash_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches cbdash dependencies to the given tag
}
