% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebase-dependencies.R
\name{add_codebase_deps}
\alias{add_codebase_deps}
\title{codebase dependencies utils}
\usage{
add_codebase_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches codebase dependencies to the given tag
}
