% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2-dependencies.R
\name{add_select2_deps}
\alias{add_select2_deps}
\title{select2 dependencies utils}
\usage{
add_select2_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches select2 dependencies to the given tag
}
