% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_button.R
\name{cb_button}
\alias{cb_button}
\title{Title}
\usage{
cb_button(
  id,
  label = NULL,
  icon = NULL,
  width = NULL,
  color = c("primary", "secondary", "success", "info", "warning", "danger"),
  size = c("md", "sm", "lg"),
  alt = FALSE,
  disabled = FALSE,
  icon_right = FALSE,
  transparent = FALSE
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
