% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_codeblock.R
\name{cb_codeblock}
\alias{cb_codeblock}
\title{Title}
\usage{
cb_codeblock(language, code)
}
\arguments{
\item{code}{}
}
\description{
Title
}
\examples{
cb_card(title='Code Example',cb_codeblock(
    language = 'r',
    'cb_codeblock <- function(language, code) {
    language <- tolower(language)
    tags$div(
        class="content content-full bg-light mb-5",
        tags$pre(
            class = "pre-sh",
            tags$code(
                class = language,
                code
            )
        )
    )
}
'
))
}
