% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_dropdown.R
\name{cb_dropdown}
\alias{cb_dropdown}
\title{Title}
\usage{
cb_dropdown(
  id,
  label = NULL,
  icon = NULL,
  actions,
  color = c("primary", "secondary", "success", "info", "warning", "danger"),
  alt = FALSE
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
