% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_elements.R
\name{cb_file_input}
\alias{cb_file_input}
\title{Title}
\usage{
cb_file_input(
  id,
  label,
  multiple = FALSE,
  accept = NULL,
  width = "100\%",
  buttonLabel = "Browse...",
  placeholder = "No file selected",
  capture = NULL,
  color = c("secondary", "info", "primary", "success", "warning", "danger")
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
