% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_header.R
\name{cb_header_dropdown}
\alias{cb_header_dropdown}
\title{Title}
\usage{
cb_header_dropdown(
  id,
  title,
  subtitle = NULL,
  options,
  icons = NULL,
  reactive_title = FALSE
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
