% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_hovercard.R
\name{cb_hovercard}
\alias{cb_hovercard}
\title{Title}
\usage{
cb_hovercard(
  id,
  title,
  subtitle = NULL,
  icon = NULL,
  bg_color = NULL,
  onclick = NULL
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
