% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_modal_button.R
\name{cb_login_dialog}
\alias{cb_login_dialog}
\title{Title}
\usage{
cb_login_dialog(id, title = NULL, ..., footer = NULL, background_color = NULL)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
