% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_modal_button.R
\name{cb_login_modal}
\alias{cb_login_modal}
\title{Title}
\usage{
cb_login_modal(
  id,
  title = NULL,
  ...,
  footer = NULL,
  brand = NULL,
  header = NULL,
  background_color = NULL
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
