% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_modal.R
\name{cb_login_news_modal}
\alias{cb_login_news_modal}
\title{Title}
\usage{
cb_login_news_modal(
  title = NULL,
  ...,
  news,
  left_footer = NULL,
  right_footer = NULL,
  brand = NULL,
  header = NULL,
  background_color = NULL,
  icon = NULL
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
