% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_login_module.R
\name{cb_login_ui}
\alias{cb_login_ui}
\title{Title}
\usage{
cb_login_ui(
  id = "login",
  brand = c("my", "app"),
  title = NULL,
  subtitle = NULL,
  news = NULL,
  contact = NULL
)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Title
}
