% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_nestable.R
\name{cb_nestable}
\alias{cb_nestable}
\title{Title}
\usage{
cb_nestable(items)
}
\arguments{
\item{items}{}
}
\description{
Title
}
\examples{
cb_nestable(
    items = list(
        'Datasets' = list(
            'dataset_1.csv',
            'dataset_2.csv'
        ),
        'Models' = list(
            'regression_model1',
            'regression_model2',
            'regression_model3'
        ),
        'Tables' = list(
            'table1',
            'table2'
        ),
        'Figures' = list(
            'figure1'
        ),
        'Plans' = list(
            'SAP1.pdf'
        ),
        'Other Documents' = list(
            'manuscript.docx'
        )
    )
)
}
