% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_content.R
\name{cb_project_block}
\alias{cb_project_block}
\title{Title}
\usage{
cb_project_block(
  ids,
  title,
  description,
  date_created,
  date_updated,
  n_datasets,
  n_models,
  n_results
)
}
\arguments{
\item{side_button}{}
}
\description{
Title
}
