% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_elements.R
\name{cb_select_input}
\alias{cb_select_input}
\title{Title}
\usage{
cb_select_input(
  id,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  select2 = TRUE,
  width = NULL,
  size = NULL,
  mb = 4
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
