% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_elements.R
\name{cb_text_input}
\alias{cb_text_input}
\title{Title}
\usage{
cb_text_input(
  id,
  label,
  value = "",
  width = NULL,
  placeholder = NULL,
  floating = FALSE,
  mb = 4
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
