% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_card.R
\name{cb_upgrade_card}
\alias{cb_upgrade_card}
\title{Title}
\usage{
cb_upgrade_card(title, cost, time, info, button, active = FALSE)
}
\arguments{
\item{button}{}
}
\description{
Title
}
\examples{
cb_upgrade_card('Developer', '$19', 'per month', list(c('2','Projects'),c('10GB Storage')),
cb_button('xx','Current Plan'))
}
