% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbdash-package.R
\docType{package}
\name{cbdash}
\alias{cbdash}
\alias{cbdash-package}
\title{cbdash}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
