% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_page.R
\name{cb_page}
\alias{cb_page}
\title{Title}
\usage{
cb_page(
  navbar,
  header = NULL,
  body,
  theme = c("none", "corporate", "earth", "elegance", "flat", "pulse")
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
